3.0-SNAPSHOT-3

+++ src/main/java/me/refracdevelopment/simplegems/api/events/impl/GemsAddEvent.java
@@ -7,9 +7,9 @@ import org.bukkit.entity.Player;
 @Getter
 public class GemsAddEvent extends SimpleEvent {
 
-    private final double newGems;
+    private final long newGems;
 
-    public GemsAddEvent(Player player, double amount) {
+    public GemsAddEvent(Player player, long amount) {
         super(player);
         this.newGems = amount;
     }
