+++ src/main/java/me/refracdevelopment/simplegems/api/events/impl/GemsPayEvent.java
@@ -7,10 +7,10 @@ import org.bukkit.entity.Player;
 @Getter
 public class GemsPayEvent extends SimpleEvent {
 
-    private final double paidGems;
+    private final long paidGems;
     private final Player target;
 
-    public GemsPayEvent(Player player, Player target, double amount) {
+    public GemsPayEvent(Player player, Player target, long amount) {
         super(player);
         this.target = target;
         this.paidGems = amount;
