+++ src/main/java/me/refracdevelopment/simplegems/data/Stat.java
@@ -2,25 +2,25 @@ package me.refracdevelopment.simplegems.data;
 
 public class Stat {
 
-    private double amount;
+    private long amount;
 
-    public double getAmount() {
+    public long getAmount() {
         return this.amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
 
-    public void incrementAmount(double amount) {
+    public void incrementAmount(long amount) {
         this.amount += amount;
     }
 
-    public void decrementAmount(double amount) {
+    public void decrementAmount(long amount) {
         this.amount -= amount;
     }
 
-    public boolean hasAmount(double amount) {
+    public boolean hasAmount(long amount) {
         return this.amount >= amount;
     }
 }
